\name{SimHaz-package}
\alias{SimHaz-package}
\alias{SimHaz}
\docType{package}
\title{
Simulated Survival and Hazard Analysis for time-dependent
}
\description{
SimHaz package in R generates power for the Cox proportional hazards model by simulating survival events data with time dependent exposure status for subjects. A dichotomous exposure variable is considered with a single transition from unexposed to exposed status during the subject time's in the study.
}
\details{
\tabular{ll}{
Package: \tab SimHaz\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2015-07-22\cr
License: \tab GPL-2 \cr
Depends: R (>= 3.1.1), survival
}
}
\author{
Danyi Xiong, Teeranan Pokaprakarn, Nusrat Rabbee \cr
Maintainer: Nusrat Rabbee <rabbee@berkeley.edu>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
